+++ src/main/java/com/mopl/api/domain/user/exception/UserErrorCode.java
@@ -1,0 +1,17 @@
+package com.mopl.api.domain.user.exception;
+
+import com.mopl.api.global.config.exception.ErrorCode;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import org.springframework.http.HttpStatus;
+
+@Getter
+@AllArgsConstructor
+public enum UserErrorCode implements ErrorCode {
+    USER_NOT_FOUND("US001", "존재하지 않는 사용자입니다.", HttpStatus.NOT_FOUND),
+    ;
+
+    private final String code;
+    private final String message;
+    private final HttpStatus httpStatus;
+}
