+++ src/test/java/com/mopl/api/domain/review/service/ReviewServiceTest.java
@@ -6,11 +6,11 @@ import com.mopl.api.domain.review.dto.request.ReviewCreateRequest;
 import com.mopl.api.domain.review.dto.request.ReviewUpdateRequest;
 import com.mopl.api.domain.review.dto.response.ReviewDto;
 import com.mopl.api.domain.review.entity.Review;
-import com.mopl.api.domain.review.exception.detail.ContentNotFoundException;
+import com.mopl.api.domain.content.exception.detail.ContentNotFoundException;
 import com.mopl.api.domain.review.exception.detail.ReviewAlreadyExistsException;
 import com.mopl.api.domain.review.exception.detail.ReviewNotFoundException;
 import com.mopl.api.domain.review.exception.detail.ReviewUnauthorizedException;
-import com.mopl.api.domain.review.exception.detail.UserNotFoundException;
+import com.mopl.api.domain.user.exception.detail.UserNotFoundException;
 import com.mopl.api.domain.review.mapper.ReviewMapper;
 import com.mopl.api.domain.review.repository.ReviewRepository;
 import com.mopl.api.domain.user.dto.response.UserDto;
